requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy[asyncio]==2.0.35
asyncpg>=0.30.0
aiosqlite==0.20.0
pydantic-settings==2.1.0
email-validator==2.1.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
